Renamed _set_functions_async -> _set_functions
